ai_model_server/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
requests